==> 23_Calculate_SPEI.ipynb <==
## Prepare Standardized Precipitation Evapotranspiration Index (SPEI)

In this notebook, the 3-month SPEI for Germany from 1981 to 2024 is calculated.
Steps include:
- Calculate the Potential Evatotranspiration from ERA5-land data (2m temperature, 10 m wind speed, surface net solar radiation and relative humidity)
- Calculate the SPEI using the water balance (total precipitation minus potential evapotranspiration)
- Crop everything to Germany and save as yearly datasets
- Also crop the SPEI downloaded from the official SPEI database to Germany and save as yearly datasets to have a comparison to the calculated values

# import libraries
import os
import xarray as xr
import geopandas as gpd
import odc.stac
import rasterio
import pyet
import numpy as np
import pandas as pd
import xclim.indicators.atmos

# set working directory
os.chdir("E:/Master/Thesis/3_Data")
print("Current working directory: {0}".format(os.getcwd()))

# prepare input datasets for calculating PET
# open era5 dataset
ds = xr.open_dataset("./Raw/ERA5-Land/data_1.nc", decode_coords="all",decode_times=True)

# rename dimensions to fit other data
ds = ds.rename({"valid_time":"time",
                            "longitude": "lon",
                            "latitude": "lat"})

# extract temperature data
t2m = ds.drop_vars(["number", "expver", "u10", "v10"])

# convert temperature from K to °C (required for pyet package)
t2m['t2m'] = t2m.t2m - 273.15

# change unit in variable attributes
t2m.t2m.attrs["units"] = "°C"

# assign crs to data
t2m = t2m.rio.write_crs('WGS 1984')

# create new variable that combines u and v component of wind
# source: https://knowledge.dea.ga.gov.au/notebooks/How_to_guides/External_data_ERA5_Climate/
ds["wind10m"] = (ds["u10"]**2 + ds["v10"]**2)**0.5

# extract wind data
wind = ds.drop_vars(["number", "expver", "u10", "v10", "t2m"])

# set unit in variable attributes
wind.wind10m.attrs["units"] = "m/s"

# assign crs to data
wind = wind.rio.write_crs('WGS 1984')

# open era5 surface net solar radiation dataset
ssr = xr.open_dataset("./Raw/ERA5-Land/data_4.nc", decode_coords="all",decode_times=True)

# rename dimensions to fit other data
ssr = ssr.rename({"valid_time":"time",
                            "longitude": "lon",
                            "latitude": "lat"})

# convert radiation from J to MJ (required for pyet package)
ssr['ssr'] = ssr.ssr / 1000000

# change unit in variable attributes
ssr.ssr.attrs["units"] = "MJ/m²"

# assign crs to data
ssr = ssr.rio.write_crs('WGS 1984')

# open era5 surface pressure dataset
sp = xr.open_dataset("./Raw/ERA5-Land/data_5.nc", decode_coords="all",decode_times=True)

# rename dimensions to fit other data
sp = sp.rename({"valid_time":"time",
                            "longitude": "lon",
                            "latitude": "lat"})

# convert pressure from Pa to kPa (required for pyet package)
sp['sp'] = sp.sp / 1000

# change unit in variable attributes
sp.sp.attrs["units"] = "kPa"

# assign crs to data
sp = sp.rio.write_crs('WGS 1984')

# open era5 2 m dewpoint temperature dataset
d2m = xr.open_dataset("./Raw/ERA5-Land/data_6.nc", decode_coords="all",decode_times=True)

# rename dimensions to fit other data
d2m = d2m.rename({"valid_time":"time",
                            "longitude": "lon",
                            "latitude": "lat"})

# convert temperature from K to °C (required for pyet package)
d2m['d2m'] = d2m.d2m - 273.15

# change unit in variable attributes
d2m.d2m.attrs["units"] = "°C"

# assign crs to data
d2m = d2m.rio.write_crs('WGS 1984')

# calculate relative humidity from mean temperature and mean dewpoint temperature
# source: equation 8 from https://journals.ametsoc.org/view/journals/bams/86/2/bams-86-2-225.xml?tab_body=pdf , dissolved for RH
# RH ≈ 100 × e^((17.625 × Td) / (243.04 + Td) - (17.625 × T) / (243.04 + T))

rh = 100 * np.exp((17.625 * d2m.d2m) / (243.04 + d2m.d2m) - (17.625 * t2m.t2m) / (243.04 + t2m.t2m))

# set unit in variable attributes
rh.attrs["units"] = "%"

# load and prepare precipitation data from era5-land
# the data has ben split into two while downloading
filelist = ("./Raw/ERA5-Land/data_2.nc", "./Raw/ERA5-Land/data_3.nc")
era5 = xr.open_mfdataset(filelist, concat_dim="valid_time", combine='nested')

# rename dimensions to fit other data
era5 = era5.rename({"valid_time":"time",
                            "longitude": "lon",
                            "latitude": "lat"})

# sort by date
era5 = era5.sortby("time")

# extract precipitation data
tp = era5.drop_vars(["number", "expver"])

# convert precipitation from "m of water equivalent per day" to "mm / month"
# source: ECMWF conversion table https://confluence.ecmwf.int/pages/viewpage.action?pageId=197702790
tp['tp'] = tp.tp * 1000 * tp.tp.time.dt.days_in_month

# change unit in variable attributes
tp.tp.attrs["units"] = "mm/month"

# assign crs to data
tp = tp.rio.write_crs('WGS 1984')

# load Germany shapefile 
ger = gpd.read_file("./GER.shp")

# extract one date to create mask of germany with dimensions of climate data (all era5-land datasets share same dimensions)
wind_ex = wind.sel(time = "1981-08-01", method = "nearest")

# create mask of research area using the dimensions of the exemplary climate data
ger_mask = rasterio.features.geometry_mask(ger.geometry, 
                                            out_shape=wind_ex.odc.geobox.shape,
                                            transform=wind_ex.odc.geobox.affine,
                                            all_touched=False,
                                            invert=False)

ger_mask = xr.DataArray(ger_mask, dims=("lat", "lon"))

# apply mask of research area to datasets
t2m_ger = t2m["t2m"].where(~ger_mask)
wind_ger = wind["wind10m"].where(~ger_mask)
ssr_ger = ssr["ssr"].where(~ger_mask)
sp_ger = sp["sp"].where(~ger_mask)
rh_ger = rh.where(~ger_mask)
tp_ger = tp["tp"].where(~ger_mask)

# calculate potential evapotranspiration 
pet = pyet.pm_fao56(tmean = t2m_ger, wind = wind_ger, rn = ssr_ger, rh = rh_ger, pressure = sp_ger)

# set unit in variable attributes
pet.attrs["units"] = "mm/month"

# calculate water balance (precipitation - potential evapotranspiration)
wb = tp_ger - pet

wb.attrs["units"] = "mm/month"

# calculate spei with xclim
spei = xclim.indicators.atmos.standardized_precipitation_evapotranspiration_index(wb = wb, 
                                                                                  freq = "MS", # monthly resampling with a 3-month moving window
                                                                                  window = 3, 
                                                                                  dist = "gamma")

spei

# save calculated spei dataset
# compress data to reduce size
comp = dict(zlib=True, complevel=4)
spei.encoding.update(comp)

# save as netCDF4 dataset
spei.to_netcdf("./SPEI/SPEI_calc_3M_GER.nc")

# for comparison, also quickly process the SPEI downloaded from the official SPEI database
spei_db = xr.open_dataset("./Raw/spei03.nc", decode_coords="all",decode_times=True)

spei_db

# drop all values before 1981
spei_db = spei_db.sel(time=slice("1981-01-01", None))

# extract one date to create mask of germany with dimensions of spei data
spei_db_ex = spei_db.sel(time = "2023-12-16", method = "nearest")

# create mask of research area using the dimensions of the exemplary data
ger_mask = rasterio.features.geometry_mask(ger.geometry, 
                                            out_shape=spei_db_ex.odc.geobox.shape,
                                            transform=spei_db_ex.odc.geobox.affine,
                                            all_touched=False,
                                            invert=False)

ger_mask = xr.DataArray(ger_mask, dims=("lat", "lon"))

# apply mask of research area to dataset
spei_db_ger = spei_db["spei"].where(~ger_mask)

# compress data to reduce size
comp = dict(zlib=True, complevel=4)
spei_db_ger.encoding.update(comp)

# save as netCDF4 dataset
spei_db_ger.to_netcdf("./SPEI/SPEI_database_3M_GER.nc")
