==> 24_Process_SPEI.ipynb <==
## Process SPEI

In this notebook, previously prepared monthly SPEI data (with a 3-month integration period) is further processed to retrieve spatially aggregated mean values per state and for Germany. For comparison, calculated SPEI values as well as SPEI values downloaded from the SPEI database are processed.

# import libraries
import os
import glob
import xarray as xr
import pandas as pd
import geopandas as gpd
import rasterio
import odc.stac
import pathlib
from pathlib import Path
import warnings

# set working directory
os.chdir("E:/Master/Thesis/3_Data")
print("Current working directory: {0}".format(os.getcwd()))

warnings.filterwarnings("ignore", category = UserWarning)
warnings.filterwarnings("ignore", category = RuntimeWarning)

# load calculated SPEI values
spei_calc = xr.open_dataset("./SPEI/SPEI_calc_3M_GER.nc",  decode_coords="all",decode_times=True)

# calculate mean values per month for Germany
spei_monthly_spatial_means = spei_calc.mean(['lon','lat'], skipna=True).to_dataframe()
spei_monthly_spatial_means

# drop unneeded columns
spei_monthly_spatial_means = spei_monthly_spatial_means.drop(['spatial_ref', 'number', 'expver'], axis=1)

# rename column with spei value to show area
spei_monthly_spatial_means.rename({"spei":"GER"}, axis=1, inplace=True)

# now, calculate monthly mean values for all federal states
# get list of shapefiles of all states
directory = "./Administrative_Borders"
files = list(pathlib.Path(directory).glob('*.shp'))

# loop over filelist to process each state
for f in range(len(files)):
    file = files[f]

    # load state shapefile 
    state = str(file)[23:25]
    state_shp = gpd.read_file(file)

    # reproject shapefile to match netcdf data
    state_shp = state_shp.to_crs(4326)

    # extract one date to create mask of state with dimensions of data
    spei_ex = spei_calc.sel(time = "1999-01-16", method = "nearest")

    # create mask of research area using the dimensions of the exemplary data
    state_mask = rasterio.features.geometry_mask(state_shp.geometry, 
                                                out_shape=spei_ex.odc.geobox.shape,
                                                transform=spei_ex.odc.geobox.affine,
                                                all_touched=False,
                                                invert=False)
    
    state_mask = xr.DataArray(state_mask, dims=("lat", "lon"))

    # mask monthly data
    spei_state = spei_calc["spei"].where(~state_mask)

    # calculate monthly mean value of state and extract to dataframe
    spei_state_spatial_means = spei_state.mean(['lon','lat'], skipna=True).to_dataframe()

    # drop unneeded columns
    spei_state_spatial_means = spei_state_spatial_means.drop(['spatial_ref', 'number', 'expver'], axis=1)

    # add monthly mean values of state to dataframe
    spei_monthly_spatial_means[state] = spei_state_spatial_means["spei"]  

spei_monthly_spatial_means

# save as csv dataset
spei_monthly_spatial_means.to_csv("./Tables/Spatial_Means/SPEI_calc_3M_Spatial_Means_GER.csv")

# now follow the same steps for the SPEI downloaded from the official database
spei_db = xr.open_dataset("./SPEI/SPEI_database_3M_GER.nc",  decode_coords="all",decode_times=True)

# calculate mean values per month for Germany
spei_monthly_spatial_means = spei_db.mean(['lon','lat'], skipna=True).to_dataframe()
spei_monthly_spatial_means

# drop unneeded columns
spei_monthly_spatial_means = spei_monthly_spatial_means.drop(['crs'], axis=1)

# rename column with spei value to show area
spei_monthly_spatial_means.rename({"spei":"GER"}, axis=1, inplace=True)

# now, calculate monthly mean values for all federal states
# get list of shapefiles of all states
directory = "./Administrative_Borders"
files = list(pathlib.Path(directory).glob('*.shp'))

# loop over filelist to process each state
for f in range(len(files)):
    file = files[f]

    # load state shapefile 
    state = str(file)[23:25]
    state_shp = gpd.read_file(file)

    # reproject shapefile to match netcdf data
    state_shp = state_shp.to_crs(4326)

    # extract one date to create mask of state with dimensions of data
    spei_ex = spei_db.sel(time = "1999-01-16", method = "nearest")

    # create mask of research area using the dimensions of the exemplary data
    state_mask = rasterio.features.geometry_mask(state_shp.geometry, 
                                                out_shape=spei_ex.odc.geobox.shape,
                                                transform=spei_ex.odc.geobox.affine,
                                                all_touched=False,
                                                invert=False)
    
    state_mask = xr.DataArray(state_mask, dims=("lat", "lon"))

    # mask monthly data
    spei_state = spei_db["spei"].where(~state_mask)

    # calculate monthly mean value of state and extract to dataframe
    spei_state_spatial_means = spei_state.mean(['lon','lat'], skipna=True).to_dataframe()

    # drop unneeded columns
    spei_state_spatial_means = spei_state_spatial_means.drop(['crs'], axis=1)

    # add monthly mean values of state to dataframe
    spei_monthly_spatial_means[state] = spei_state_spatial_means["spei"]  

spei_monthly_spatial_means

# save as csv dataset
spei_monthly_spatial_means.to_csv("./Tables/Spatial_Means/SPEI_database_3M_Spatial_Means_GER.csv")
