==> 32_Process_Calculated_GPP.ipynb <==
## Process Calculated GPP

In this notebook, previously calculated GPP data NIRv is further processed to retrieve spatially aggregated monthly mean values for Germany and federal states.

# import libraries
import os
import glob
import xarray as xr
import pandas as pd
import geopandas as gpd
import rasterio
import odc.stac
import pathlib
from pathlib import Path
import warnings

# set working directory
os.chdir("E:/Master/Thesis/3_Data")
print("Current working directory: {0}".format(os.getcwd()))

warnings.filterwarnings("ignore", category = UserWarning)
warnings.filterwarnings("ignore", category = RuntimeWarning)

# spatially aggregate monthly GPP data
# to get mean values for Germany and federal states

# first, get values for broadleaved forest
# load all monthly gpp values as one dataset
directory = f"./GPP/GPP_NIRv/Monthly"
filelist = glob.glob(os.path.join(directory, "*BL*.nc"))
gpp_monthly = xr.open_mfdataset(filelist, concat_dim="time", combine="nested", engine="netcdf4")

# calculate mean gpp values per month for Germany
gpp_monthly_spatial_means = gpp_monthly.mean(['lon','lat'], skipna=True).to_dataframe()
gpp_monthly_spatial_means

# drop unneeded columns
gpp_monthly_spatial_means = gpp_monthly_spatial_means.drop(['crs', 'spatial_ref'], axis=1)

# rename column with gpp value to show area
gpp_monthly_spatial_means.rename({"GPP":"GER"}, axis=1, inplace=True)

# now, calculate monthly mean gpp values for all federal states
# get list of shapefiles of all states
directory = "./Administrative_Borders"
files = list(pathlib.Path(directory).glob('*.shp'))

# loop over filelist to process each state
for f in range(len(files)):
    file = files[f]

    # load state shapefile 
    state = str(file)[23:25]
    state_shp = gpd.read_file(file)

    # reproject shapefile to match netcdf data
    state_shp = state_shp.to_crs(4326)

    # extract one date to create mask of state with dimensions of data
    gpp_ex = gpp_monthly.sel(time = "2000-01-01", method = "nearest")

    # create mask of research area using the dimensions of the exemplary data
    state_mask = rasterio.features.geometry_mask(state_shp.geometry, 
                                                out_shape=gpp_ex.odc.geobox.shape,
                                                transform=gpp_ex.odc.geobox.affine,
                                                all_touched=False,
                                                invert=False)
    
    state_mask = xr.DataArray(state_mask, dims=("lat", "lon"))

    # mask data
    gpp_state = gpp_monthly["GPP"].where(~state_mask)

    # calculate monthly mean gpp of state and extract to dataframe
    gpp_state_spatial_means = gpp_state.mean(['lon','lat'], skipna=True).to_dataframe()

    # drop unneeded columns
    gpp_state_spatial_means = gpp_state_spatial_means.drop(['crs', 'spatial_ref'], axis=1)

    # add monthly mean gpp values of state to dataframe
    gpp_monthly_spatial_means[state] = gpp_state_spatial_means["GPP"]   

gpp_monthly_spatial_means

# save as csv dataset
gpp_monthly_spatial_means.to_csv("./Tables/Spatial_Means/GPP_NIRv_Monthly_Spatial_Means_BL.csv")

# now follow the same steps for needleleaved forest
# load all monthly gpp values as one dataset
directory = f"./GPP/GPP_NIRv/Monthly"
filelist = glob.glob(os.path.join(directory, "*NL*.nc"))
gpp_monthly = xr.open_mfdataset(filelist, concat_dim="time", combine="nested", engine="netcdf4")

# calculate mean gpp values per month for Germany
gpp_monthly_spatial_means = gpp_monthly.mean(['lon','lat'], skipna=True).to_dataframe()
gpp_monthly_spatial_means

# drop unneeded columns
gpp_monthly_spatial_means = gpp_monthly_spatial_means.drop(['crs', 'spatial_ref'], axis=1)

# rename column with gpp value to show area
gpp_monthly_spatial_means.rename({"GPP":"GER"}, axis=1, inplace=True)

# now, calculate monthly mean gpp values for all federal states
# get list of shapefiles of all states
directory = "./Administrative_Borders"
files = list(pathlib.Path(directory).glob('*.shp'))

# loop over filelist to process each state
for f in range(len(files)):
    file = files[f]

    # load state shapefile 
    state = str(file)[23:25]
    state_shp = gpd.read_file(file)

    # reproject shapefile to match netcdf data
    state_shp = state_shp.to_crs(4326)

    # extract one date to create mask of state with dimensions of data
    gpp_ex = gpp_monthly.sel(time = "2000-01-01", method = "nearest")

    # create mask of research area using the dimensions of the exemplary data
    state_mask = rasterio.features.geometry_mask(state_shp.geometry, 
                                                out_shape=gpp_ex.odc.geobox.shape,
                                                transform=gpp_ex.odc.geobox.affine,
                                                all_touched=False,
                                                invert=False)
    
    state_mask = xr.DataArray(state_mask, dims=("lat", "lon"))

    # mask data
    gpp_state = gpp_monthly["GPP"].where(~state_mask)

    # calculate monthly mean gpp of state and extract to dataframe
    gpp_state_spatial_means = gpp_state.mean(['lon','lat'], skipna=True).to_dataframe()

    # drop unneeded columns
    gpp_state_spatial_means = gpp_state_spatial_means.drop(['crs', 'spatial_ref'], axis=1)

    # add monthly mean gpp values of state to dataframe
    gpp_monthly_spatial_means[state] = gpp_state_spatial_means["GPP"]   

gpp_monthly_spatial_means

# save as csv dataset
gpp_monthly_spatial_means.to_csv("./Tables/Spatial_Means/GPP_NIRv_Monthly_Spatial_Means_NL.csv")
