==> 35_Calculate_Overall_Means.ipynb <==
## Calculate overall means

In this notebook, mean values over the full observation period of all analyzed variables are calculated for the federal states and Germany. The resulting table will give a quick overview of mean values / dimensions of variables in the different states.

# import libraries
import os
import pandas as pd
import re

# set working directory
os.chdir("E:/Master/Thesis/3_Data")
print("Current working directory: {0}".format(os.getcwd()))

# filenames with spatial means for states
file_names = [
    'NDVI_Yearly_Spatial_Means_BL_harmonized.csv',
    'NDVI_Yearly_Spatial_Means_NL_harmonized.csv',
    'LAI_Yearly_Spatial_Means_BL_harmonized.csv',
    'LAI_Yearly_Spatial_Means_NL_harmonized.csv',
    'FAPAR_Yearly_Spatial_Means_BL_harmonized.csv',
    'FAPAR_Yearly_Spatial_Means_NL_harmonized.csv',
    'Temperature_Yearly_Spatial_Means_BL.csv',
    'Temperature_Yearly_Spatial_Means_NL.csv',
    'Precipitation_Yearly_Spatial_Means_GER.csv',
    'Windspeed_Yearly_Spatial_Means_BL.csv',
    'Windspeed_Yearly_Spatial_Means_NL.csv',
    'Soil_Water_1_Yearly_Spatial_Means_BL.csv',
    'Soil_Water_1_Yearly_Spatial_Means_NL.csv',
    'Soil_Water_2_Yearly_Spatial_Means_BL.csv',
    'Soil_Water_2_Yearly_Spatial_Means_NL.csv',
    'Soil_Water_3_Yearly_Spatial_Means_BL.csv',
    'Soil_Water_3_Yearly_Spatial_Means_NL.csv',
    'Soil_Water_4_Yearly_Spatial_Means_BL.csv',
    'Soil_Water_4_Yearly_Spatial_Means_NL.csv',
    'SPEI_calc_3M_Spatial_Means_GER.csv',
    'GPP_NIRv_Monthly_Spatial_Means_BL_harmonized.csv',
    'GPP_NIRv_Monthly_Spatial_Means_NL_harmonized.csv',
    'NPP_calc_Monthly_Spatial_Means_BL.csv',
    'NPP_calc_Monthly_Spatial_Means_NL.csv'
]

# states
state_columns = ['GER', 'BB', 'BE', 'BW', 'BY', 'HB', 'HE', 'HH', 'MV', 'NI', 'NW', 'RP', 'SH', 'SL', 'SN', 'ST', 'TH']

# list to store means for each variable
all_overall_means = []

for file_name in file_names:
        
    # open the csv file 
    df = pd.read_csv(f'./Tables/Spatial_Means/{file_name}')

    # extract base variable name and region
    base_variable_name = ""
    region = ""

    # remove endings
    temp_name = file_name.replace('.csv', '').replace('_harmonized', '')

    # extract region (BL, NL, GER) from the end of the filename string
    region_match = re.search(r'_(BL|NL|GER)$', temp_name)
    region = region_match.group(1)

    # remove the region from temp_name to isolate the variable part more easily
    temp_name_without_region = temp_name[:-len(region_match.group(0))]

    # extract the base variable name 
    variable_match = re.search(r'(.+)_(Yearly|Monthly)_Spatial_Means', temp_name_without_region)
    if variable_match:
        base_variable_name = variable_match.group(1)
    else:
        # for the case of SPEI, where there is no 'yearly' or 'monthly' in filename
        base_variable_name = 'SPEI'


    # combine base variable name and region
    variable_name_with_region = f"{base_variable_name} {region}"

    # get state columns
    present_state_columns = [col for col in state_columns if col in df.columns]

    # calculate the mean for each present state column over the entire dataset
    overall_mean_series = df[present_state_columns].mean()

    # convert the series to a dictionary and add the variable name
    mean_dict = overall_mean_series.to_dict()
    mean_dict['Variable'] = variable_name_with_region
    all_overall_means.append(mean_dict)

# create a dataframe from the collected overall means
combined_overall_means_df = pd.DataFrame(all_overall_means)

combined_overall_means_df.head()

# reorder columns to have 'Variable' as the first column
cols = ['Variable'] + [col for col in combined_overall_means_df.columns if col != 'Variable']
combined_overall_means_df = combined_overall_means_df[cols]

combined_overall_means_df

# save the combined table
combined_overall_means_df.to_csv('./Tables/Spatial_Means/Combined_Overall_Spatial_Means.csv', index=False)
